+++ heat/core/manipulations.py
@@ -1732,7 +1732,7 @@ def repeat(a: Iterable, repeats: Iterable, axis: Optional[int] = None) -> DNDarr
     return repeated_array
 
 
-def reshape(a: DNDarray, shape: Tuple[int, ...], new_split: Optional[int] = None) -> DNDarray:
+def reshape(a: DNDarray, *shape: Union[int, Tuple[int, ...]], **kwargs) -> DNDarray:
     """
     Returns an array with the same data and number of elements as `a`, but with the specified shape.
 
@@ -1740,7 +1740,7 @@ def reshape(a: DNDarray, shape: Tuple[int, ...], new_split: Optional[int] = None
     ----------
     a : DNDarray
         The input array
-    shape : Tuple[int,...]
+    shape : Union[int, Tuple[int,...]]
         Shape of the new array
     new_split : int, optional
         The new split axis if `a` is a split DNDarray. None denotes same axis.
@@ -1771,7 +1771,39 @@ def reshape(a: DNDarray, shape: Tuple[int, ...], new_split: Optional[int] = None
     if not isinstance(a, DNDarray):
         raise TypeError("'a' must be a DNDarray, currently {}".format(type(a)))
 
+    # check for nested sequence
+    if not np.isscalar(shape[0]):
+        shape = shape[0]
+    if shape == -1:
+        shape = (a.gnumel,)
+
     tdtype, tdevice = a.dtype.torch_type(), a.device.torch_device
+
+    # check for valid shape type, shape size
+    a_proxy = torch.ones((1,)).as_strided(a.gshape, [0] * a.ndim)
+    try:
+        a_proxy.reshape(shape)
+    except TypeError as e:
+        # allow array-like `shape`
+        try:
+            shape = shape.tolist()
+        except AttributeError:
+            raise TypeError(e)
+    except (RuntimeError, ValueError):
+        shape = list(shape)
+        shape_size = torch.prod(torch.tensor(shape, dtype=torch.int, device=tdevice)).item()
+        if shape.count(-1) > 1:
+            raise ValueError("too many unknown dimensions")
+        elif shape.count(-1) == 1:
+            if a.size % shape_size != 0:
+                raise ValueError(
+                    "cannot reshape array of size {} into shape {}".format(a.size, shape)
+                )
+        else:
+            if a.size != shape_size:
+                raise ValueError(
+                    "cannot reshape array of size {} into shape {}".format(a.size, shape)
+                )
 
     def reshape_argsort_counts_displs(
         shape1, lshape1, displs1, axis1, shape2, displs2, axis2, comm
@@ -1808,38 +1840,43 @@ def reshape(a: DNDarray, shape: Tuple[int, ...], new_split: Optional[int] = None
         displs[1:] = torch.cumsum(counts[:-1], dim=0)
         return argsort, counts, displs
 
-    if shape == -1:
-        shape = (a.gnumel,)
-
-    if not isinstance(shape, (list, tuple)):
-        raise TypeError("shape must be list, tuple, currently {}".format(type(shape)))
-
     # check new_split parameter
+    new_split = kwargs.get("new_split")
     if new_split is None:
         new_split = a.split
-    stride_tricks.sanitize_axis(shape, new_split)
-
-    # Check the type of shape and number elements
-    shape = stride_tricks.sanitize_shape(shape, -1)
+    new_split = stride_tricks.sanitize_axis(shape, new_split)
 
     shape = list(shape)
     shape_size = torch.prod(torch.tensor(shape, dtype=torch.int, device=tdevice))
 
-    # infer unknown dimension
-    if shape.count(-1) > 1:
-        raise ValueError("too many unknown dimensions")
-    elif shape.count(-1) == 1:
+    if shape.count(-1) == 1:
         pos = shape.index(-1)
         shape[pos] = int(-(a.size / shape_size).item())
         shape_size *= -shape[pos]
 
-    if shape_size != a.size:
-        raise ValueError("cannot reshape array of size {} into shape {}".format(a.size, shape))
-
     # Forward to Pytorch directly
     if a.split is None:
-        return factories.array(
-            torch.reshape(a.larray, shape), dtype=a.dtype, device=a.device, comm=a.comm
+        local_reshape = torch.reshape(a.larray, shape)
+        if new_split is None:
+            return DNDarray(
+                local_reshape,
+                tuple(shape),
+                dtype=a.dtype,
+                split=None,
+                device=a.device,
+                comm=a.comm,
+                balanced=True,
+            )
+        _, _, local_slice = a.comm.chunk(shape, new_split)
+        local_reshape = local_reshape[local_slice]
+        return DNDarray(
+            local_reshape,
+            tuple(shape),
+            dtype=a.dtype,
+            split=new_split,
+            comm=a.comm,
+            device=a.device,
+            balanced=True,
         )
 
     # Create new flat result tensor
@@ -1860,7 +1897,7 @@ def reshape(a: DNDarray, shape: Tuple[int, ...], new_split: Optional[int] = None
         shape, local_shape, new_displs, new_split, a.shape, old_displs, a.split, a.comm
     )
 
-    # rearange order
+    # rearrange order
     send = a.larray.flatten()[sendsort]
     a.comm.Alltoallv((send, sendcounts, senddispls), (data, recvcounts, recvdispls))
 
@@ -1871,10 +1908,18 @@ def reshape(a: DNDarray, shape: Tuple[int, ...], new_split: Optional[int] = None
     # Reshape local tensor
     data = data.reshape(local_shape)
 
-    return factories.array(data, dtype=a.dtype, is_split=new_split, device=a.device, comm=a.comm)
+    return DNDarray(
+        data,
+        tuple(shape),
+        dtype=a.dtype,
+        split=new_split,
+        device=a.device,
+        comm=a.comm,
+        balanced=True,
+    )
 
 
-DNDarray.reshape = lambda self, shape, axis=None: reshape(self, shape, axis)
+DNDarray.reshape = lambda self, *shape, **kwargs: reshape(self, *shape, **kwargs)
 DNDarray.reshape.__doc__ = reshape.__doc__